move string extensions
